+++ examples/all-clusters-app/ameba/main/include/LevelControlCommands.h
@@ -515,8 +515,7 @@ CHIP_ERROR LevelControlReadCurrentLevel(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::CurrentLevel::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -527,8 +526,7 @@ CHIP_ERROR LevelControlReadRemainingTime(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::RemainingTime::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -539,8 +537,7 @@ CHIP_ERROR LevelControlReadMinLevel(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::MinLevel::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -551,8 +548,7 @@ CHIP_ERROR LevelControlReadMaxLevel(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::MaxLevel::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -563,8 +559,7 @@ CHIP_ERROR LevelControlReadCurrentFrequency(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::CurrentFrequency::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -575,8 +570,7 @@ CHIP_ERROR LevelControlReadMinFrequency(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::MinFrequency::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -587,8 +581,7 @@ CHIP_ERROR LevelControlReadMaxFrequency(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::MaxFrequency::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -599,8 +592,7 @@ CHIP_ERROR LevelControlReadOptions(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::Options::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -611,8 +603,7 @@ CHIP_ERROR LevelControlReadOnOffTransitionTime(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::OnOffTransitionTime::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -623,8 +614,7 @@ CHIP_ERROR LevelControlReadOnLevel(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::OnLevel::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -635,8 +625,7 @@ CHIP_ERROR LevelControlReadOnTransitionTime(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::OnTransitionTime::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -647,8 +636,7 @@ CHIP_ERROR LevelControlReadOffTransitionTime(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::OffTransitionTime::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -659,8 +647,7 @@ CHIP_ERROR LevelControlReadDefaultMoveRate(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::DefaultMoveRate::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
@@ -671,8 +658,7 @@ CHIP_ERROR LevelControlReadStartUpCurrentLevel(int argc, char ** argv)
     data->attributeId         = Clusters::LevelControl::Attributes::StartUpCurrentLevel::Id;
     data->clusterId           = Clusters::LevelControl::Id;
     data->isReadAttribute     = true;
-    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", static_cast<uint32_t>(data->clusterId),
-                    static_cast<uint32_t>(data->attributeId));
+    ChipLogProgress(NotSpecified, "Read cluster=0x%x, attribute=0x%08x", data->clusterId, data->attributeId);
     DeviceLayer::PlatformMgr().ScheduleWork(SwitchWorkerFunction, reinterpret_cast<intptr_t>(data));
     return CHIP_NO_ERROR;
 }
